Stop GetWindowAttribute truncating the window tile data pointer
WINDOW_TILE_DATA is a pointer, but the accessor was declared to return u32, so
the truncation happened in the return type itself -- all ten call sites
dutifully cast the result back to a pointer, but the top half was already gone.
The bag's pocket names were blank because PrintPocketNames copies through it.

Widened to uintptr_t rather than patching the call sites: one change covers all
ten, and it is a no-op on the GBA target where uintptr_t is already 32 bits. One
caller stored the result in a u32 local first, so that is widened too.

Worth noting how convincingly this hid: VRAM_ is allocated around 0x1060xxxxx,
so a truncated VRAM pointer reads as 0x0600xxxx -- which looks exactly like a
real GBA VRAM address and invites the conclusion that the value is fine.

+++ src/window.c
@@ -555,7 +555,7 @@ bool8 SetWindowAttribute(u8 windowId, u8 attributeId, u32 value)
     }
 }
 
-u32 GetWindowAttribute(u8 windowId, u8 attributeId)
+uintptr_t GetWindowAttribute(u8 windowId, u8 attributeId)
 {
     switch (attributeId)
     {
@@ -574,7 +574,11 @@ u32 GetWindowAttribute(u8 windowId, u8 attributeId)
     case WINDOW_BASE_BLOCK:
         return gWindows[windowId].window.baseBlock;
     case WINDOW_TILE_DATA:
-        return (u32)(gWindows[windowId].tileData);
+        // Returned through uintptr_t, not u32: this attribute is a pointer, and
+        // a u32 return truncated it on a 64-bit host. Ten call sites cast the
+        // result straight back to a pointer -- PrintPocketNames among them,
+        // which is why the bag's pocket names were blank.
+        return (uintptr_t)(gWindows[windowId].tileData);
     default:
         return 0;
     }
